Mount TMPDIR inside the sandbox
systemd-repart copies files into a directory in TMPDIR
before running mksquashfs.

mkosi sets TMPDIR to the workspace directory to keep its tempfiles
organised.

When running in a tools tree sandbox it mounts / as a tmpfs
and the root directory of the build image on top at its usual path.
This root directory is inside the workspace directory,
but the workspace directory itself is not mounted.

This results in repart copying files into a tmpfs,
which could cause problems for large trees,
but also results in extended attributes being lost
since tmpfs doesn't support them.

This in turn results in it being impossible to create
non-strict system extension images
since the extended attribute to enable this gets lost.