Set extra groups in INVOKING_USER.mkdir
If a parent directory of the path to create isn't traversable
then mkdir will fail to create the directory even if the immediate parent
exists and has the correct permissions.

This can happen in shared development hardware where storage is allocated
using a directory with a common unix group e.g.

$ ls -ld /data
drwxrws--- 4 root devgroup 4096 Mar 12 17:11 /data