Add RuntimeBuildSources= setting
This setting mounts the build sources and build directory into a
booted container/virtual machine in the same location that they were
mounted to do the build.

This helps both with making gdb more useful in the container as the
sources will be available for use by gdb without having to do any
source directory mapping. It also helps with doing incremental builds
inside the container.