Fix systemd-journal-remote termination trampling exit status
The non-exceptional terminate would propagate into an exception
if the command were terminated directly and left to check the exit code.

By adding a flag to terminate on context exit to spawn() we can propagate an
exception if the process exited early and only raise an exception
if it didn't terminate after being sent the signal.