Add option to support systemd-sysupdate naming style
When using `systemd-sysupdate` the names of output files need to contain the
image version since this is the only version that is reasonable to use across
all artifacts. This change makes sure the UKI stored in the ESP or XBOOTLDR
partition has an appropriate name, along with the name if `SplitArtifacts=`
is set.

Note: The UKI cannot contain the kernel version or other values that will
change and cannot be hard-coded into a `sysupdate.d` file.