Add `UnifiedKernelImageName=`
This can be used to control the name to use for the UKI during image
generation. This only is supported for when a single UKI is used.

This is useful for the `systemd-sysupdate` case, as you can set this to
`%i_%v` to use a format that can be parse by its configuration. The
current format used includes both a roothash as well as the kernel
version which both can't be matched by sysupdate.