opensuse: Add support for aarch64
This requires another rework for the repositories() method for
OpenSUSE. Whereas before we picked up all the repositories from
https://download.opensuse.org/tumbleweed/repo/, it turns out that
the debug and source repositories can also be found at
https://download.opensuse.org/debug/tumbleweed/repo/ and
https://download.opensuse.org/source/tumbleweed/repo/ respectively.
Furthermore, the latter locations are the only ones that are available
for ports under https://download.opensuse.org/ports/ so we switch to
those instead.

Additionally, openSUSE-current is not available for ports so we disallow
using current, stable and release with architectures other than x86-64.

While there are more ports than just aarch64, for now let's just add
aarch64 and wait for user demand before we add any others.